# Use a worker-specific random source to remove lock contention.