--- a/statsd/worker_test.go
+++ b/statsd/worker_test.go
@@ -10,10 +10,13 @@ import (
 func TestShouldSample(t *testing.T) {
 	rates := []float64{0.01, 0.05, 0.1, 0.25, 0.5, 0.75, 0.9, 0.99, 1.0}
 	iterations := 50_000
-	worker := newWorker(newBufferPool(1, 1, 1), nil)
+
 	for _, rate := range rates {
 		rate := rate // Capture range variable.
 		t.Run(fmt.Sprintf("Rate %0.2f", rate), func(t *testing.T) {
+			t.Parallel()
+
+			worker := newWorker(newBufferPool(1, 1, 1), nil)
 			count := 0
 			for i := 0; i < iterations; i++ {
 				if worker.shouldSample(rate) {