### What does this PR do ?
Use a separate random number source per worker. Currently, the `worker` function `shouldSample()` calls `rand.Float64()`, which uses a lock to serialize all calls to the package-global pseudorandom number generator, creating lock contention between all instances of `worker`.

Changes:
- Create and seed a separate pseudorandom number generator for every `worker` to remove the lock contention https://github.com/DataDog/datadog-go/issues/80
- Add a test for the `shouldSample()` function to assert that it returns `true` the expected percentage of the time
- Add a benchmark for the `shouldSample()` function to measure improvement from removing lock contention

### Benchmarks against master
```
name            old time/op  new time/op  delta
ShouldSample     444ns ± 0%   190ns ± 1%  -57.26%  (p=0.000 n=8+9)
ShouldSample-2   876ns ± 2%    94ns ± 0%  -89.27%  (p=0.000 n=10+9)
ShouldSample-4  1.17µs ± 1%  0.09µs ± 2%  -92.59%  (p=0.000 n=8+8)
ShouldSample-8  1.25µs ± 3%  0.09µs ± 4%  -92.77%  (p=0.000 n=9+10)
```